Fix assume-filename handling in clang-format.el
When 'buffer-file-name' is nil 'call-process-region' returned returned
a segmentation fault error.

This was a problem when using clang-format-buffer on an orgmode source
code editing buffer.

I fixed this problem by excluding the '-assume-filename' argument
when 'buffer-file-name' is nil.

To make it a bit more flexible I also added an optional argument,
'assume-file', to specify an assume-filename that overrides
'buffer-file-name'.